requirements.txt
hydra-core==1.3.2
flwr==1.4.0
keras==2.12.0
numpy==1.23.5
omegaconf==2.3.0
tensorflow==2.12.0
mlflow==2.5.0
boto3==1.28.16
tensorflow-privacy==0.8.10
paho-mqtt==1.6.1
